+++ Runtime/Scripts/LsEventDispatcher.cs
@@ -1,0 +1,17 @@
+﻿using UnityEngine;
+using UnityEngine.Events;
+
+namespace Eos.UxKit
+{
+    public class LsEventDispatcher : MonoBehaviour
+    {
+        [SerializeField] private string _eventID;
+
+        public UnityAction<string> onEvent = delegate {};
+
+        public void DispatchEvent()
+        {
+            onEvent.Invoke(_eventID);
+        }
+    }
+}